+++ src/app/azureResourceIdentifier/models/azureResourceIdentifier.ts
