+++ src/app/azureResourceIdentifier/models/azureResourceIdentifierQuery.ts
