+++ src/app/azureResourceIdentifier/models/azureResourceIdentifierQueryResult.ts
