+++ src/app/azureResourceIdentifier/models/azureResourceIdentifierType.ts
