+++ src/app/azureResourceIdentifier/models/azureResourceManagementEndpoint.ts
