+++ src/app/azureResourceIdentifier/services/azureResourceIdentifierService.spec.ts
