+++ src/app/azureResourceIdentifier/services/azureResourceIdentifierService.ts
