+++ src/app/azureResourceIdentifier/services/azureSubscriptionService.spec.ts
